chore: move maps to pkg
